Feature #230: rename tools into lib
